requirements.txt
streamlit==1.33.0
pandas==2.2.2
requests==2.32.3
plotly==5.20.0
python-dateutil==2.9.0.post0
